requirements.txt
streamlit==1.33.0
torch==2.0.1+cpu
torchvision==0.15.2+cpu
torchaudio==2.0.2+cpu
opencv-python-headless
numpy
Pillow
matplotlib
pyyaml
--extra-index-url https://download.pytorch.org/whl/cpu